refactor: i improved the Value Objects made
